final edit to devices
Still struggling to get a control of connected devices and disconnect a connected device. Can get a connection going and a good status going, but nothing once it is connected already. Switching over to use a more simplistic and hacky method of changing between connected and disconnected each time a device is tapped